+++ programs/portal/src/state.rs
@@ -31,10 +31,10 @@ impl PortalGlobal {
     pub fn update_index(&mut self, message_id: [u8; 32], new_index: u128) {
         self.m_index = max(new_index, self.m_index);
 
-         emit!(MTokenIndexReceived {
+        emit!(MTokenIndexReceived {
             index: new_index,
             message_id,
-        }); 
+        });
     }
 
     pub fn generate_message_id(&mut self) -> [u8; 32] {
@@ -63,6 +63,6 @@ impl BridgeMessage {
 
 #[event]
 pub struct MTokenIndexReceived {
-    pub index: u64,
+    pub index: u128,
     pub message_id: [u8; 32],
 }
